services/auth/pyproject.toml
[tool.poetry]
name = "grimoire-auth-service"
version = "0.1.0"
description = "Grimoire Authentication Microservice"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pydantic = "^2.0.0"
sqlalchemy = "^2.0.0"
pydantic-settings = "^2.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
sqlmodel = "^0.0.12"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
email-validator = "^2.0.0"
httpx = "^0.24.1"
python-dotenv = "^1.0.0"
confluent-kafka = "^2.3.0"
aiokafka = "^0.8.1"
avro-python3 = "^1.10.2"
fastavro = "^1.9.0"
prometheus-client = "^0.18.0"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
protobuf = "^4.24.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
ruff = "^0.0.287"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
